requirements.txt
streamlit
supabase
python-dotenv
librosa
numpy
scipy
soundfile
praat-parselmouth
pandas
scikit-learn
requests
